requirements.txt
django==4.0.4
python-dotenv==0.20.0
djongo==1.3.6
pytz==2022.1
pymongo==3.12.3
djangorestframework==3.13.1
pandas==1.4.2